# Speed up Erlang 500x by using nodelay and system_time